move test_w
