+++ astropy/cosmology/tests/test_cosmology.py
@@ -97,74 +97,6 @@ def test_distance_broadcast():
             assert allclose(value_flat, value_3d.flatten())
 
 
-# This class is to test whether the routines work correctly
-# if one only overloads w(z)
-class test_cos_sub(flrw.FLRW):
-    def __init__(self):
-        super().__init__(70.0, 0.27, 0.73, Tcmb0=0.0, name="test_cos")
-        self._w0 = -0.9
-
-    def w(self, z):
-        return self._w0 * np.ones_like(z)
-
-
-# Similar, but with neutrinos
-class test_cos_subnu(flrw.FLRW):
-    def __init__(self):
-        super().__init__(
-            70.0, 0.27, 0.73, Tcmb0=3.0, m_nu=0.1 * u.eV, name="test_cos_nu"
-        )
-        self._w0 = -0.8
-
-    def w(self, z):
-        return self._w0 * np.ones_like(z)
-
-
-@pytest.mark.skipif(not HAS_SCIPY, reason="test requires scipy")
-def test_de_subclass():
-    # This is the comparison object
-    z = [0.2, 0.4, 0.6, 0.9]
-    cosmo = flrw.wCDM(H0=70, Om0=0.27, Ode0=0.73, w0=-0.9, Tcmb0=0.0)
-    # Values taken from Ned Wrights advanced cosmo calculator, Aug 17 2012
-    assert allclose(
-        cosmo.luminosity_distance(z), [975.5, 2158.2, 3507.3, 5773.1] * u.Mpc, rtol=1e-3
-    )
-    # Now try the subclass that only gives w(z)
-    cosmo = test_cos_sub()
-    assert allclose(
-        cosmo.luminosity_distance(z), [975.5, 2158.2, 3507.3, 5773.1] * u.Mpc, rtol=1e-3
-    )
-    # Test efunc
-    assert allclose(cosmo.efunc(1.0), 1.7489240754, rtol=1e-5)
-    assert allclose(cosmo.efunc([0.5, 1.0]), [1.31744953, 1.7489240754], rtol=1e-5)
-    assert allclose(cosmo.inv_efunc([0.5, 1.0]), [0.75904236, 0.57178011], rtol=1e-5)
-    # Test de_density_scale
-    assert allclose(cosmo.de_density_scale(1.0), 1.23114444, rtol=1e-4)
-    assert allclose(
-        cosmo.de_density_scale([0.5, 1.0]), [1.12934694, 1.23114444], rtol=1e-4
-    )
-
-    # Add neutrinos for efunc, inv_efunc
-
-
-@pytest.mark.skipif(not HAS_SCIPY, reason="test requires scipy")
-def test_efunc_vs_invefunc_flrw():
-    """Test that efunc and inv_efunc give inverse values"""
-    z0 = 0.5
-    z = np.array([0.5, 1.0, 2.0, 5.0])
-
-    # FLRW is abstract, so requires test_cos_sub defined earlier
-    # This requires scipy, unlike the built-ins, because it
-    # calls de_density_scale, which has an integral in it
-    cosmo = test_cos_sub()
-    assert allclose(cosmo.efunc(z0), 1.0 / cosmo.inv_efunc(z0))
-    assert allclose(cosmo.efunc(z), 1.0 / cosmo.inv_efunc(z))
-    # Add neutrinos
-    cosmo = test_cos_subnu()
-    assert allclose(cosmo.efunc(z0), 1.0 / cosmo.inv_efunc(z0))
-    assert allclose(cosmo.efunc(z), 1.0 / cosmo.inv_efunc(z))
-
-
 @pytest.mark.skipif(not HAS_SCIPY, reason="test requires scipy")
 def test_distances():
     # Test distance calculations for various special case
